+++ docs-src/pages/UiAutocomplete.vue
@@ -557,6 +557,13 @@
 
                         <tbody>
                             <tr>
+                                <td><code>focus()</code></td>
+                                <td>
+                                    <p>Call this method to programmatically focus the autocomplete.</p>
+                                </td>
+                            </tr>
+
+                            <tr>
                                 <td><code>reset()</code></td>
                                 <td>
                                     <p>Call this method to reset the autocomplete to its initial value. You should also reset the <code>invalid</code> and <code>touched</code> props.</p>