+++ docs-src/pages/UiFileupload.vue
@@ -267,6 +267,29 @@
                     </table>
                 </div>
             </ui-tab>
+
+            <ui-tab title="Methods">
+                <div class="table-responsive">
+                    <table class="table">
+                        <thead>
+                            <tr>
+                                <th>Name</th>
+                                <th>Description</th>
+                            </tr>
+                        </thead>
+
+                        <tbody>
+                            <tr>
+                                <td><code>openPicker()</code></td>
+                                <td>
+                                    <p>Call this method to programmatically open the file picker.</p>
+                                    <p>This method should only be called in response to a <a href="https://github.com/WICG/interventions/issues/12" target="_blank" rel="noopener">user gesture</a> like a click or focus or the action may fail.</p>
+                                </td>
+                            </tr>
+                        </tbody>
+                    </table>
+                </div>
+            </ui-tab>
         </ui-tabs>
     </section>
 </template>